fixed delete

+++ nutri-journal/controllers/api/weightLogs.js
@@ -50,12 +50,12 @@ const getByDate = async (req, res) => {
 
 const deleteWeight = async (req, res) => {
   try {
-    const { date } = req.query;
+    const { date, userId } = req.query;
 
-    const existingLog = await WeightLog.findOne({ date: new Date(date) });
+    const existingLog = await WeightLog.findOne({ date: new Date(date), userId: userId });
 
     if (existingLog) {
-      await WeightLog.deleteOne({ date: new Date(date) });
+      await WeightLog.deleteOne({ date: new Date(date), userId: userId });
       res.status(200).json({ deleted: true });
     } else {
       res.status(200).json({ deleted: false });
